pyproject.toml
[project]
name = "pydjimqtt"
version = "0.1.0"
description = "DJI Cloud API Python SDK"
readme = "README.md"
authors = [
    { name = "GrooveWJH", email = "groovewjh@foxmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "markdown-it-py>=4.0.0",
    "mdurl>=0.1.2",
    "paho-mqtt>=2.1.0",
    "pygments>=2.19.2",
    "pynput>=1.8.1",
    "pyobjc>=12.1",
    "rich>=14.2.0",
    "six>=1.17.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pydjimqtt"]
